HEEDLS-439 Rename web CustomPromptHelper class to CentreCustomPromptHelper
