+++ DigitalLearningSolutions.Web/Controllers/Register/RegisterDelegateByCentreController.cs
@@ -27,8 +27,8 @@ namespace DigitalLearningSolutions.Web.Controllers.Register
     public class RegisterDelegateByCentreController : Controller
     {
         private const string CookieName = "DelegateRegistrationByCentreData";
+        private readonly CentreCustomPromptHelper centreCustomPromptHelper;
         private readonly ICryptoService cryptoService;
-        private readonly CustomPromptHelper customPromptHelper;
         private readonly IJobGroupsDataService jobGroupsDataService;
         private readonly IRegistrationService registrationService;
         private readonly IUserDataService userDataService;
@@ -37,7 +37,7 @@ namespace DigitalLearningSolutions.Web.Controllers.Register
         public RegisterDelegateByCentreController(
             IJobGroupsDataService jobGroupsDataService,
             IUserService userService,
-            CustomPromptHelper customPromptHelper,
+            CentreCustomPromptHelper centreCustomPromptHelper,
             ICryptoService cryptoService,
             IUserDataService userDataService,
             IRegistrationService registrationService
@@ -45,7 +45,7 @@ namespace DigitalLearningSolutions.Web.Controllers.Register
         {
             this.jobGroupsDataService = jobGroupsDataService;
             this.userService = userService;
-            this.customPromptHelper = customPromptHelper;
+            this.centreCustomPromptHelper = centreCustomPromptHelper;
             this.userDataService = userDataService;
             this.registrationService = registrationService;
             this.cryptoService = cryptoService;
@@ -114,7 +114,7 @@ namespace DigitalLearningSolutions.Web.Controllers.Register
 
             var centreId = data.Centre!.Value;
 
-            customPromptHelper.ValidateCustomPrompts(
+            centreCustomPromptHelper.ValidateCustomPrompts(
                 centreId,
                 model.Answer1,
                 model.Answer2,
@@ -303,7 +303,7 @@ namespace DigitalLearningSolutions.Web.Controllers.Register
             int centreId
         )
         {
-            return customPromptHelper.GetEditCustomFieldViewModelsForCentre(
+            return centreCustomPromptHelper.GetEditCustomFieldViewModelsForCentre(
                 centreId,
                 model.Answer1,
                 model.Answer2,
@@ -316,7 +316,7 @@ namespace DigitalLearningSolutions.Web.Controllers.Register
 
         private IEnumerable<CustomFieldViewModel> GetCustomFieldsFromData(DelegateRegistrationData data)
         {
-            return customPromptHelper.GetCustomFieldViewModelsForCentre(
+            return centreCustomPromptHelper.GetCustomFieldViewModelsForCentre(
                 data.Centre!.Value,
                 data.Answer1,
                 data.Answer2,
