+++ DigitalLearningSolutions.Web/Startup.cs
@@ -171,7 +171,6 @@ namespace DigitalLearningSolutions.Web
             services.AddScoped<IUserService, UserService>();
             services.AddScoped<IUserDataService, UserDataService>();
             services.AddScoped<ICryptoService, CryptoService>();
-            services.AddScoped<ICustomPromptsService, CustomPromptsService>();
             services.AddScoped<ICentreCustomPromptsService, CentreCustomPromptsService>();
             services.AddScoped<ICourseAdminFieldsService, CourseAdminFieldsService>();
             services.AddScoped<ICentreCustomPromptsDataService, CentreCustomPromptsDataService>();
@@ -184,7 +183,7 @@ namespace DigitalLearningSolutions.Web
             services.AddScoped<IPasswordService, PasswordService>();
             services.AddScoped<IPasswordDataService, PasswordDataService>();
             services.AddScoped<IDelegateApprovalsService, DelegateApprovalsService>();
-            services.AddScoped<CustomPromptHelper>();
+            services.AddScoped<CentreCustomPromptHelper>();
             services.AddScoped<IClockService, ClockService>();
             services.AddScoped<ISupportTicketDataService, SupportTicketDataService>();
             services.AddScoped<IRoleProfileService, RoleProfileService>();
